+++ build.gradle
@@ -28,11 +28,11 @@ test {
 }
 
 application {
-    mainClass.set("seedu.duke.Duke")
+    mainClass.set("Main")
 }
 
 shadowJar {
-    archiveBaseName = "duke"
+    archiveBaseName = "BobAndSteve"
     archiveClassifier = null
 }
