--- src/main/java/bobAndSteve/command/CommandEnum.java
@@ -1,13 +1,0 @@
-package bobAndSteve.command;
-
-public enum CommandEnum {
-    HELP,
-    MARK,
-    UNMARK,
-    TODO,
-    DEADLINE,
-    EVENT,
-    LIST,
-    DELETE,
-    BYE;
-}
