--- src/main/java/bobAndSteve/exception/InvalidCommandException.java
@@ -1,7 +1,0 @@
-package bobAndSteve.exception;
-
-public class InvalidCommandException extends BobAndSteveException {
-    public InvalidCommandException(String message) {
-        super(message);
-    }
-}
