--- src/main/java/bobAndSteve/exception/InvalidCommandFormatException.java
@@ -1,7 +1,0 @@
-package bobAndSteve.exception;
-
-public class InvalidCommandFormatException extends BobAndSteveException {
-    public InvalidCommandFormatException(String message) {
-        super(message);
-    }
-}
