+++ src/main/java/bobandsteve/command/CommandEnum.java
@@ -1,0 +1,19 @@
+package bobandsteve.command;
+
+/**
+ * Enum representing the available commands in the application.
+ * Each constant corresponds to a specific command that can be executed
+ * by the program.
+ */
+public enum CommandEnum {
+    HELP,
+    MARK,
+    UNMARK,
+    TODO,
+    DEADLINE,
+    EVENT,
+    LIST,
+    DELETE,
+    FIND,
+    BYE;
+}
