+++ src/test/java/bobandsteve/command/CommandTest.java
@@ -3,9 +3,9 @@ package bobandsteve.command;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
 
-import bobAndSteve.command.*;
-import bobAndSteve.exception.InvalidCommandFormatException;
 import org.junit.jupiter.api.Test;
+
+import bobandsteve.exception.InvalidCommandFormatException;
 
 public class CommandTest {
     @Test
